backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
openai-whisper==20231117
TTS==0.20.0
yt-dlp==2023.11.16
deep-translator==1.11.4
pydub==0.25.1
numpy==1.24.3
torch==2.1.0
torchaudio==2.1.0
soundfile==0.12.1
scipy==1.11.4
moviepy==1.0.3
ffmpeg-python==0.2.0
aiofiles==23.2.1
